+++ src/main/java/com/condeinsyt/skuulba/controller/AuthController.java
@@ -1,7 +1,7 @@
 package com.condeinsyt.skuulba.controller;
 
 import com.condeinsyt.skuulba.model.User;
-import com.condeinsyt.skuulba.service.AuthServiceImpl;
+import com.condeinsyt.skuulba.service.impl.AuthServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
