+++ src/main/java/com/condeinsyt/skuulba/controller/StudentHealthController.java
@@ -2,7 +2,7 @@ package com.condeinsyt.skuulba.controller;
 
 
 import com.condeinsyt.skuulba.model.StudentHealth;
-import com.condeinsyt.skuulba.service.StudentHealthService;
+import com.condeinsyt.skuulba.service.interfaces.StudentHealthService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
