+++ src/main/java/com/condeinsyt/skuulba/controller/StudentPersonalityController.java
@@ -2,7 +2,7 @@ package com.condeinsyt.skuulba.controller;
 
 import com.condeinsyt.skuulba.dto.StudentPersonalityDTO;
 import com.condeinsyt.skuulba.model.StudentPersonality;
-import com.condeinsyt.skuulba.service.StudentPersonalityService;
+import com.condeinsyt.skuulba.service.interfaces.StudentPersonalityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
