+++ src/main/java/com/condeinsyt/skuulba/controller/StudentPhysicalController.java
@@ -1,10 +1,8 @@
 package com.condeinsyt.skuulba.controller;
 
 import com.condeinsyt.skuulba.dto.StudentPhysicalDTO;
-import com.condeinsyt.skuulba.model.ArrayHolder;
 import com.condeinsyt.skuulba.model.StudentPhysical;
-import com.condeinsyt.skuulba.service.StudentPhysicalService;
-import com.condeinsyt.skuulba.service.StudentPhysicalServiceImpl;
+import com.condeinsyt.skuulba.service.interfaces.StudentPhysicalService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
