+++ src/main/java/com/condeinsyt/skuulba/service/impl/StudentAttendanceServiceImpl.java
@@ -1,17 +1,16 @@
-package com.condeinsyt.skuulba.service;
+package com.condeinsyt.skuulba.service.impl;
 
 import com.condeinsyt.skuulba.dto.StudentAttendanceDTO;
 import com.condeinsyt.skuulba.model.Student;
 import com.condeinsyt.skuulba.model.StudentAttendance;
 import com.condeinsyt.skuulba.repository.StudentAttendanceRepository;
+import com.condeinsyt.skuulba.service.interfaces.StudentAttendanceService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class StudentAttendanceServiceImpl implements StudentAttendanceService {
