+++ src/main/java/com/condeinsyt/skuulba/service/impl/StudentHandicapServiceImpl.java
@@ -1,8 +1,9 @@
-package com.condeinsyt.skuulba.service;
+package com.condeinsyt.skuulba.service.impl;
 
 import com.condeinsyt.skuulba.dto.StudentHandicapDTO;
 import com.condeinsyt.skuulba.model.StudentHandicap;
 import com.condeinsyt.skuulba.repository.StudentHandicapRepository;
+import com.condeinsyt.skuulba.service.interfaces.StudentHandicapService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
