+++ src/main/java/com/condeinsyt/skuulba/service/impl/StudentPhysicalServiceImpl.java
@@ -1,8 +1,9 @@
-package com.condeinsyt.skuulba.service;
+package com.condeinsyt.skuulba.service.impl;
 
-import com.condeinsyt.skuulba.dto.StudentEducationDTO;
-import com.condeinsyt.skuulba.model.StudentEducation;
-import com.condeinsyt.skuulba.repository.StudentEducationRepository;
+import com.condeinsyt.skuulba.dto.StudentPhysicalDTO;
+import com.condeinsyt.skuulba.service.interfaces.StudentPhysicalService;
+import com.condeinsyt.skuulba.model.StudentPhysical;
+import com.condeinsyt.skuulba.repository.StudentPhysicalRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -10,29 +11,27 @@ import org.springframework.stereotype.Service;
 import java.util.*;
 
 @Service
-public class StudentEducationServiceImpl implements StudentEducationService {
+public class StudentPhysicalServiceImpl implements StudentPhysicalService {
 
-    private StudentEducationRepository studentEducationRepository;
-
+    private StudentPhysicalRepository studentPhysicalRepository;
 
     @Autowired
-    public StudentEducationServiceImpl(StudentEducationRepository studentEducationRepository){
-        this.studentEducationRepository = studentEducationRepository;
+    public StudentPhysicalServiceImpl(StudentPhysicalRepository studentPhysicalRepository){
+        this.studentPhysicalRepository = studentPhysicalRepository;
     }
 
-
     @Override
-    public HashMap<String, Object> createEducation(StudentEducationDTO studentEducationDTO) {
+    public HashMap<String, Object> createPhysical(StudentPhysicalDTO studentPhysical) {
         HashMap<String, Object> responseData = new HashMap<>();
 
-        ArrayList<StudentEducation> content = studentEducationDTO.getEducations();
+        ArrayList<StudentPhysical> content = studentPhysical.getPhysicals();
 
         try{
-            for(StudentEducation studentE : content){
-                StudentEducation newStudentEducation  = this.studentEducationRepository.save(studentE);
+            for(StudentPhysical studentP : content){
+                StudentPhysical newStudentPhysical  = this.studentPhysicalRepository.save(studentP);
             }
             responseData.put("data",null);
-            responseData.put("message","Education Added to student");
+            responseData.put("message","Physical Added to student");
             responseData.put("status", HttpStatus.OK.value());
         }catch (Exception e){
             e.printStackTrace();
@@ -47,22 +46,22 @@ public class StudentEducationServiceImpl implements StudentEducationService {
     }
 
     @Override
-    public HashMap<String, Object> updateEducation(StudentEducation studentEducation) {
+    public HashMap<String, Object> updatePhysical(StudentPhysical studentPhysical) {
         HashMap<String, Object> responseData = new HashMap<>();
 
         try{
-            Optional<StudentEducation> studentEducationFound = studentEducationRepository.findById(studentEducation.getId());
+            Optional<StudentPhysical> studentPhysicalFound = studentPhysicalRepository.findById(studentPhysical.getId());
 
-            if(!studentEducationFound.isPresent()){
+            if(!studentPhysicalFound.isPresent()){
                 responseData.put("data", Collections.EMPTY_LIST);
-                responseData.put("message","No student education found");
+                responseData.put("message","No student physique found");
                 responseData.put("status",HttpStatus.NO_CONTENT.value());
                 return responseData;
             }
 
-            StudentEducation updatedEducation  = studentEducationRepository.save(studentEducation);
-            responseData.put("data",updatedEducation);
-            responseData.put("message","Student Education updated successfully");
+            StudentPhysical updatedStudent  = studentPhysicalRepository.save(studentPhysical);
+            responseData.put("data",updatedStudent);
+            responseData.put("message","Student updated successfully");
             responseData.put("status",HttpStatus.OK.value());
         }catch (Exception e){
             e.printStackTrace();
@@ -75,23 +74,22 @@ public class StudentEducationServiceImpl implements StudentEducationService {
     }
 
     @Override
-    public HashMap<String, Object> deleteEducation(Long id) {
+    public HashMap<String, Object> deletePhysical(Long id) {
         HashMap<String, Object > responseData = new HashMap<>();
 
         try{
 
-            Optional<StudentEducation> studentEducationFound = studentEducationRepository.findById(id);
-            Long studId = studentEducationFound.get().getStudentId();
+            Optional<StudentPhysical> studentPhysicalFound = studentPhysicalRepository.findById(id);
+            Long studId = studentPhysicalFound.get().getStudentId();
 
-
-//            if(!studentEducationFound.isPresent()){
-//                responseData.put("data", Collections.EMPTY_LIST);
-//                responseData.put("message","Student Education not found \uD83E\uDD7A");
-//                responseData.put("status",HttpStatus.NO_CONTENT.value());
-//                return responseData;
-//            }
-            studentEducationRepository.deleteById(id);
-            return findStudentEducationById(studId);
+            if(!studentPhysicalFound.isPresent()){
+                responseData.put("data", Collections.EMPTY_LIST);
+                responseData.put("message","Student Physical not found \uD83E\uDD7A");
+                responseData.put("status",HttpStatus.NO_CONTENT.value());
+                return responseData;
+            }
+            studentPhysicalRepository.deleteById(id);
+           return listStudentPhysical(studId);
         } catch(Exception e){
             e.printStackTrace();
             responseData.put("data",null);
@@ -103,17 +101,17 @@ public class StudentEducationServiceImpl implements StudentEducationService {
     }
 
     @Override
-    public HashMap<String, Object> listStudentEducation(Long id) {
+    public HashMap<String, Object> listStudentPhysical(Long id) {
         HashMap<String, Object> responseData = new HashMap<>();
         try {
-            List<StudentEducation> studentEducationList = studentEducationRepository.findAllByStudentId(id);
-            if (studentEducationList.isEmpty()) {
+            List<StudentPhysical> studentPhysicalList = studentPhysicalRepository.findAllByStudentId(id);
+            if (studentPhysicalList.isEmpty()) {
                 responseData.put("data", Collections.EMPTY_LIST);
-                responseData.put("message", "Student education not found \uD83E\uDD7A");
+                responseData.put("message", "Student physique not found \uD83E\uDD7A");
                 responseData.put("status", HttpStatus.NO_CONTENT.value());
                 return responseData;
             }
-            responseData.put("data", studentEducationList);
+            responseData.put("data", studentPhysicalList);
             responseData.put("message", "Student physique found \uD83D\uDE42");
             responseData.put("status", HttpStatus.OK.value());
         } catch (Exception e) {
@@ -127,18 +125,18 @@ public class StudentEducationServiceImpl implements StudentEducationService {
     }
 
     @Override
-    public HashMap<String, Object> findStudentEducationById(Long id) {
+    public HashMap<String, Object> findStudentPhysicalById(Long id) {
         HashMap<String, Object> responseData = new HashMap<>();
 
-        Optional<StudentEducation> foundStudentEducation = studentEducationRepository.findById(id);
+        Optional<StudentPhysical> foundStudentPhysical = studentPhysicalRepository.findById(id);
         try{
-            if(!foundStudentEducation.isPresent()){
+            if(!foundStudentPhysical.isPresent()){
                 responseData.put("data", Collections.EMPTY_LIST);
-                responseData.put("message","Student Education not found \uD83E\uDD7A");
+                responseData.put("message","Student Physical not found \uD83E\uDD7A");
                 responseData.put("status",HttpStatus.NO_CONTENT.value());
                 return responseData;
             }
-            responseData.put("data",foundStudentEducation);
+            responseData.put("data",foundStudentPhysical);
             responseData.put("message","Student Physical found \uD83D\uDE42");
             responseData.put("status",HttpStatus.OK.value());
         }catch (Exception e){
