+++ src/main/java/com/condeinsyt/skuulba/service/impl/StudentTalentServiceImpl.java
@@ -1,8 +1,9 @@
-package com.condeinsyt.skuulba.service;
+package com.condeinsyt.skuulba.service.impl;
 
 import com.condeinsyt.skuulba.dto.StudentTalentsDTO;
 import com.condeinsyt.skuulba.model.StudentTalent;
 import com.condeinsyt.skuulba.repository.StudentTalentRepository;
+import com.condeinsyt.skuulba.service.interfaces.StudentTalentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
