+++ src/main/java/com/condeinsyt/skuulba/service/impl/UserServiceImpl.java
@@ -1,7 +1,8 @@
-package com.condeinsyt.skuulba.service;
+package com.condeinsyt.skuulba.service.impl;
 
 import com.condeinsyt.skuulba.model.User;
 import com.condeinsyt.skuulba.repository.UserRepository;
+import com.condeinsyt.skuulba.service.interfaces.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
