+++ src/main/java/com/condeinsyt/skuulba/service/interfaces/StudentAttendanceService.java
@@ -1,4 +1,4 @@
-package com.condeinsyt.skuulba.service;
+package com.condeinsyt.skuulba.service.interfaces;
 
 import com.condeinsyt.skuulba.dto.StudentAttendanceDTO;
 import com.condeinsyt.skuulba.model.Student;
