+++ src/main/java/com/condeinsyt/skuulba/service/interfaces/StudentHandicapService.java
@@ -1,4 +1,4 @@
-package com.condeinsyt.skuulba.service;
+package com.condeinsyt.skuulba.service.interfaces;
 
 import com.condeinsyt.skuulba.dto.StudentHandicapDTO;
 import com.condeinsyt.skuulba.model.StudentHandicap;
