+++ src/main/java/com/condeinsyt/skuulba/service/interfaces/StudentPhysicalService.java
@@ -1,4 +1,4 @@
-package com.condeinsyt.skuulba.service;
+package com.condeinsyt.skuulba.service.interfaces;
 
 import com.condeinsyt.skuulba.dto.StudentPhysicalDTO;
 import com.condeinsyt.skuulba.model.ArrayHolder;
