+++ src/main/java/com/condeinsyt/skuulba/service/interfaces/StudentTalentService.java
@@ -1,4 +1,4 @@
-package com.condeinsyt.skuulba.service;
+package com.condeinsyt.skuulba.service.interfaces;
 
 import com.condeinsyt.skuulba.dto.StudentTalentsDTO;
 import com.condeinsyt.skuulba.model.StudentTalent;
